Move Action model to osf.models
