+++ osf/models/__init__.py
@@ -35,3 +35,4 @@ from osf.models.admin_profile import AdminProfile  # noqa
 from osf.models.admin_log_entry import AdminLogEntry  # noqa
 from osf.models.maintenance_state import MaintenanceState  # noqa
 from osf.models.quickfiles import QuickFilesNode  # noqa
+from osf.models.action import Action  # noqa
