+++ osf/models/action.py
@@ -5,19 +5,19 @@ from django.db import models
 
 from include import IncludeManager
 
-from osf.models.base import ObjectIDMixin
-from osf.utils.fields import NonNaiveDateTimeField
-
 from reviews.workflow import Triggers
 from reviews.workflow import States
 
+from osf.models.base import BaseModel, ObjectIDMixin
+from osf.utils.fields import NonNaiveDateTimeField
 
-class Action(ObjectIDMixin, models.Model):
+
+class Action(ObjectIDMixin, BaseModel):
 
     objects = IncludeManager()
 
-    target = models.ForeignKey('osf.PreprintService', related_name='actions')
-    creator = models.ForeignKey('osf.OSFUser', related_name='+')
+    target = models.ForeignKey('PreprintService', related_name='actions')
+    creator = models.ForeignKey('OSFUser', related_name='+')
 
     trigger = models.CharField(max_length=31, choices=Triggers.choices())
     from_state = models.CharField(max_length=31, choices=States.choices())
